+++ vuepress/docs/shared/changelog.md
@@ -47,3 +47,4 @@
 - v2.2.0
   - Add clear filters callback to options-header slot.
   - Define slot types and add to documentation.
+  - Fix column shown state in ColumnSelectionItem.